Backend Big Update

+++ practice_figma/backend/controllers/classController.js
@@ -1,5 +1,11 @@
 const Class = require('../models/classModel')
 const Student = require('../models/studentModel')
+const ClassSubject = require('../models/classSubjectModel')
+const mongoose = require('mongoose')
+const {
+    createClassSubjects,
+    deleteClassSubjectsByClassID
+} = require('../controllers/classSubjectController')
 
 const createClass = async (req, res) => {
     const {name, gradeLevel} = req.body
@@ -17,7 +23,22 @@ const createClass = async (req, res) => {
     }
 
     try {
+        if (!/^[A-Z]\d{1,2}$/.test(name)) {
+            throw Error('Class name must match the pattern /^[A-Z]\\d{1,2}$/ (e.g., A1, B12).')
+        }
+
+        if (![10, 11, 12].includes(gradeLevel)) {
+            throw Error('gradeLevel must be one of 10, 11, 12.');
+        }
+
         const new_class = await Class.create({ name: name, gradeLevel: gradeLevel })
+
+        const classsubject = await createClassSubjects(new_class._id)
+
+        if (!classsubject.success) {
+            res.status(400).json({ error: classsubject.error })
+        }
+
         res.status(200).json(new_class)
     } catch (error) {
         res.status(400).json({ error: error.message })
@@ -27,32 +48,79 @@ const createClass = async (req, res) => {
 const deleteClass = async (req, res) => {
     const {id} = req.params
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'No such class'})
+    }
+
     const deleted_class = await Class.findByIdAndDelete({ _id: id})
 
     if (!deleted_class) {
         return res.status(400).json({ error: 'No such class' })
     }
-    return res.status(200).json(deleteClass)
+
+    await Student.updateMany({ ClassID: id }, { $unset: { ClassID: "" } })
+
+    const classsubject = await deleteClassSubjectsByClassID(id)
+
+    if (!classsubject.success) {
+        return res.status(400).json({ error: classsubject.error })
+    }
+
+    return res.status(200).json(deleted_class)
 }
 
 const updateClass = async (req, res) => {
     const {id} = req.params
+    const {name, gradeLevel, numofStudents} = req.body
 
-    const updated_class = await Class.findById(id)
-
-    if (!updated_class) {
-        return res.status(400).json({error: 'No such class'})
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'No such class'})
     }
 
-    // Count the current number of students in the class
-    const currentStudentCount = await Student.countDocuments({ ClassID: id })
-
-    if (req.body.numofStudents < currentStudentCount) {
-        return res.status(400).json({ error: `New maxStudents value must be higher than or equal to the current number of students (${currentStudentCount})` })
+    if (name) {
+        if (!/^[A-Z]\d{1,2}$/.test(name)) {
+            return res.status(400).json({ error: 'Class name must match the pattern /^[A-Z]\\d{1,2}$/ (e.g., A1, B12).' });
+        }
+    }
+    
+    if (gradeLevel) {
+        if (![10, 11, 12].includes(gradeLevel)) {
+            return res.status(400).json({ error: 'gradeLevel must be one of 10, 11, 12.' });
+        }
     }
 
-    updated_class.numofStudents = req.body.numofStudents
-    await updated_class.save()
+    if (numofStudents) {
+        return res.status(400).json({ error: 'Changing this is not allowed' })
+    }
+
+    const update_class = Class.findById(id)
+
+    if (!update_class) {
+        return res.status(400).json({ error: 'No such class' })
+    }
+
+    if (name && gradeLevel) {
+        const classExists = await Class.exists({ name, gradeLevel })
+        if (classExists) {
+            return res.status(400).json({ error: 'Class with this name and grade level exists' })
+        }
+    } else if (name) {
+        const exist_gradeLevel = update_class.gradeLevel
+        const classExists = await Class.exists({ name, exist_gradeLevel })
+        if (classExists) {
+            return res.status(400).json({ error: 'Class with this name and grade level exists' })
+        }
+    } else if (gradeLevel) {
+        const exist_name = update_class.name
+        const classExists = await Class.exists({ exist_name, gradeLevel })
+        if (classExists) {
+            return res.status(400).json({ error: 'Class with this name and grade level exists' })
+        }
+    }
+
+    const updated_class = await Class.findOneAndUpdate({ _id: id }, {
+        ...req.body
+    }, { new: true })
 
     res.status(200).json(updated_class)
 }
@@ -66,8 +134,28 @@ const getClasses = async (req, res) => {
 const getClass = async (req, res) => {
     const {id} = req.params
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({error: 'No such class'})
+    }
+
     const get_class = await Class.findOne({ _id: id })
     return res.status(200).json(get_class)
+}
+
+const getClassesByTeacherID = async (req, res) => {
+    const TeacherID = req.params.TeacherID
+
+    const classsubjects = await ClassSubject.find({ TeacherID: TeacherID }).select('ClassID -_id')
+ 
+    if (!classsubjects.length) {
+        return res.status(404).json({ error: 'No classes found for this teacher' })
+    }
+
+    const classIDs = classsubjects.map(cs => cs.ClassID)
+
+    const classes = await Class.find({ _id: { $in: classIDs } })
+
+    res.status(200).json(classes)
 }
 
 module.exports = {
@@ -75,5 +163,6 @@ module.exports = {
     getClass,
     getClasses,
     updateClass,
-    deleteClass
+    deleteClass,
+    getClassesByTeacherID
 }
